Testmap lighting research
After researching the LWRP and testing out its capabilities, I've found that it seems too limited to support what we want out of this game. Rolling back to the default render pipeline or upgrading to HDRP might be our best bets. Perhaps ask John about the technical implications of both